Extend solr client to support subqueries
Solr allows to run sub queries per document using a document
transformer. For this, a synthetic field needs to be added to the
field list of the main query and a `:[subquery]` appended to it, e.g.
`creator:[subquery]`. Then the actual sub query can be added to the
`params` section of the main query - prefixed by the field name. The
sub query works exactly like a main query, the parameters need only to
be prefixed so solr can do the query and associate the result
correctly. Solr will execute that query and adds the results as a
value to the defined field. The result is the complete `ResponseBody`,
including the header. So even if there are no results, there will be a
response object (containing an empty `docs` array).

The new case class `SubQuery` allows to specify such a sub query - it
looks very similar to `QueryData`, but doesn't define another `params`
and also doesn't define a `facet` parameter, because it is not needed.
When adding a sub query to a query, the queries `params` object is
updated accordingly and the new field is added to the field list. This
then requires to explicitly define a field list for the main query,
otherwise only the synthetic field will be returned as it is the only
one in the field list after a `SubQuery` has been added.

Solr-docs: https://solr.apache.org/guide/solr/latest/query-guide/document-transformers.html#subquery
Issue: #146