Additionally mount search endpoint to /api/search/query
This little change is meant as preparation to mount the search
endpoint in a more flexible way and to finally allow to also provide a
`/version` endpoint.

The idea is that we remove the `stripPrefix` in gateway and pass the
request as-is to the search service. The current public path should
not change, therefore it stays as `/api/search` and the current
"search" endpoint will be `/api/search/query`. Later a version
`/api/search/version` and potentially other public endpoints can be
provided as siblings. Currently it is not possible to provide more
endpoints without editing the gateway configuration (besides putting
everything under `/search` which is too strange).

This commit allows clients (UI) to use the new path for doing search
queries, while the old path is enabled as well.