Return details for a project namespace
The solr query is amended with another subquery to resolve existing
`namespace` fields in the search results. This is only interesting for
projects, as these entities use that field denoting "being in that
namespace". `Group` and `User` entities on the other hand use the
`namespace` field to *define* this namespace. So the sub-query is
constraint to only `User` or `Group` entities. Solr applies this to
all result documents, but it is only decoded for projects. `User` and
`Group` entities resolve to themselves. So we potentially transfer
lots of data from solr only to discard it, sadly. I couldn't find a
way around it with the current model.

It would be nice to have a clear distinction between defining the
namespace and being a member of it (using different fields), but
unfortunately we decided against that in the past.

The search http api *replaces* the value of the `namespace` field with
the full object of either `User` or `Group` in its results, making
this a breaking change wrt the search api result structure.

Since SOLR cannot guarantee to have a `User` or `Group` being resolved
given a namespace, the result is defined as optional. When this
happens an error is logged and `None` is returned as a result.