WIP: start copy-data cmd
This command is supposed to connect to postgres and solr to copy the
data from postgres to solr. This is meant as a emergency escape if the
index is completely broken. It intentionally circumvents redis (and
other things in between). This should be quite easy to do. Only
problem is the authzed component that I cannot just query postgres
for. It requires to send RPC requests, but I couldn't figure out how
to connect to it (getting certificate errors). At least as far as I
understand it.