Add endpoint to trigger re-creating the index from the stream
Sometimes it is necessary to re-create the entire index from the raw
data (for example, when solr schema changes happen). This is done by
reading the redis stream from the beginning (or from a given message).

The endpoint is only available from the provisioning process and not
from the api.