Protect schema migration from running in parallel
Running schema migrations must not happen in parallel, should there be
multiple instance deployed. It is now protected using the
`DocumentLockResource`.

Since this changes the version-document used to track the current
schema version, this document must now be converted as a first step to
make it up to date on existing deployments.