Schema migration run only if not applied
The solr schema can be fetched and it is used to test each command in
the schema migration if it had already been applied. If not, the
command is run and otherwise it is ignored.

With this change it is possible to drop solr entirely, including
schema information. On restart, the schema will be fixed by running
those changes that haven't been applied.