Reset lock documents on startup (#211)
In the situation the services crashes or gets killed while a lock is
being held, it would startup into an incorrect state. On startup, the
lock documents are resetted to recover from such a case.

This can be problematic when many services are deployed, because it
would reset all documents right now. This will be addressed in a
future change and for the currently foreseeable future, there will be
only one instance.