Refactor config loading, preparing for writing documentation
This refactoring allows to track what config values are used by which
service. The `ConfigValues` class tracks accessing the definied config
keys by adding them to a map. So only the config keys required by a
configuration loader are then possible to retrieve from this map.

This will be used to generate the used environment variables for each
service (provision and api).

This changes the envrionment variables for queue names! They were
written in camel-case which is not very idiomatic, now are all
uppercase as it is known for env variables.