Integrating Sentry into search-api and provision services (#218)
There are these additional config values:

```
RS_SENTRY_DSN=<the dsn for search service>
RS_SENTRY_ENV=dev | prod
RS_SENTRY_ENABLED=true|false
```
The first two settings are only read, if `RS_SENTRY_ENABLED` is `true`. When this is configured, the services send events to sentry on certain points. By default everything logged as error or warn is sent. There are also a few events that are sent: when the app starts up and when a reprovision-started message is received.